[wip] move install over
