+++ lib/construct/index.js
@@ -511,7 +511,7 @@ class Constructor extends EventEmitter {
     // MAKE ASYNC/AWAIT (await x 4)
     async.series({
       unpack: this.packer.unpack({ content, installPath, statusWriter }),
-      install: this.install.bind(this, spec, installPath, statusWriter),
+      install: this.packer.install({ spec, installPath, statusWriter }),
       pack: this.packer.pack({ pkgDir, tarball, statusWriter }),
       upload: this.packer.upload({ spec, tarball, statusWriter })
     }, (err) => {
@@ -552,38 +552,6 @@ class Constructor extends EventEmitter {
       app.contextLog.info('Uploaded tarball for package', assign({ url }, logOpts));
       next();
     });
-  }
-
-  /**
-   * Install the dependencies of the package with npm.
-   * Uses the provided environment.
-   *
-   * @param {Object} spec Spec
-   * @param {String} installPath os.tmpdir base path to run the install in.
-   * @param {StatusWriter} statusWriter The writer for the status-api
-   * @param {Function} fn Completion callback.
-   * @api public
-   */
-  install(spec, installPath, statusWriter, fn) {
-    const pkgDir = path.join(installPath, 'package');
-
-    const done = fn || function () {};
-    const statusKey = 'npm install-all';
-
-    statusWriter.writeStart(statusKey);
-    const op = retry.op(this.retry);
-    // MAKE ASYNC/AWAIT
-    // https://www.npmjs.com/package/retryme#asyncawait-support
-    op.attempt(next => {
-      npm.install({
-        log: this.app.contextLog,
-        userconfig: this.app.npmrc,
-        installPath,
-        pkgDir,
-        spec,
-        statusWriter
-      }, next);
-    }, statusWriter.writeWrap(statusKey, done));
   }
 
   /**
