+++ lib/construct/npm/index.js
@@ -10,7 +10,7 @@ const assign = Object.assign;
 const statusKey = 'npm install';
 
 // MAKE ASYNC/AWAIT
-exports.install = function (opts, callback) {
+function install (opts, callback) {
   const { log, spec, installPath, pkgDir, userconfig, statusWriter } = opts;
   const done = once(callback);
   const env = assign({}, process.env);
@@ -68,5 +68,15 @@ exports.install = function (opts, callback) {
     statusWriter.write(statusKey, `'npm install' attempt completed successfully`);
     return done();
   });
-
 };
+
+function installAsync (opts) {
+  return new Promise((resolve, reject) => {
+    install(opts, e => e ? reject(err) : resolve()
+  });
+}
+
+exports = {
+  install,
+  installAsync
+}
