+++ lib/construct/packer.js
@@ -1,6 +1,9 @@
 const through = require('through2');
 const once = require('one-time');
+const retry = require('retryme');
 const zlib = require('zlib');
+const path = require('path');
+const npm = require('./npm');
 const tar = require('tar');
 
 class Packer {
@@ -46,8 +49,47 @@ class Packer {
     }
   }
 
-  install() {
+  /**
+   * Install the dependencies of the package with npm.
+   * Uses the provided environment.
+   *
+   * @param {Object} options
+   * @param {Object} options.spec Spec
+   * @param {String} options.installPath os.tmpdir base path to run the install in.
+   * @param {StatusWriter} options.statusWriter The writer for the status-api
+   *
+   * @returns {Promise} completion handler
+   */
+   async _install({ spec, installPath, statusWriter }) {
+     const pkgDir = path.join(installPath, 'package');
+     const statusKey = 'npm install-all';
 
+     statusWriter.writeStart(statusKey);
+     const op = retry.op(this.retry);
+     // MAKE ASYNC/AWAIT
+     // https://www.npmjs.com/package/retryme#asyncawait-support
+     await op.async(() => npm.installAsync({
+       log: this.app.contextLog,
+       userconfig: this.app.npmrc,
+       installPath,
+       pkgDir,
+       spec,
+       statusWriter
+     }));
+
+     return new Promise((resolve, reject) => {
+       statusWriter.writeWrap(statusKey, (err, data) => {
+         err ? reject(err) : resolve(data);
+       });
+     });
+   }
+
+  install(options) {
+    return next => {
+      this._install(options)
+        .then(next)
+        .catch(next);
+    }
   }
 
   /**
